hotfix: 매칭테스트 결과 ssr로 변경, 카카오톡 공유 키 오류 해결
