--- .env
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_SERVER_URL='https://abcode.shop/api/v1'
