+++ app/login/page.tsx
@@ -7,7 +7,7 @@ const page = () => {
   const host = headers().get('host');
 
   const redirectUri = `${protocol}://${host}/login/kakao`;
-  const apiKey = process.env.KAKAO_REST_API_KEY;
+  const apiKey = process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY;
   const kakaoLoginLink = `https://kauth.kakao.com/oauth/authorize?client_id=${apiKey}&redirect_uri=${redirectUri}&response_type=code`;
 
   return (
