+++ components/matchingtest/ResultModal.tsx
@@ -71,7 +71,7 @@ const ResultModal = ({ onClose }: ResultModalProps) => {
   const kakaoShare = () => {
     if (window.Kakao) {
       const kakao = window.Kakao;
-      const appKey = process.env.NEXT_PUBLIC_KAKAO_APP_KEY;
+      const appKey = process.env.NEXT_PUBLIC_KAKAO_JAVASCRIPT_KEY;
       const templateId = 103843;
       if (!kakao.isInitialized()) {
         kakao.init(appKey);
@@ -84,6 +84,7 @@ const ResultModal = ({ onClose }: ResultModalProps) => {
       });
     }
   };
+
   return (
     <div
       className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50"
